BackEnd/MedBotAssist.BotOpenIA/requirements.txt
beautifulsoup4
langgraph 
langchain[openai]
langchain-community
chromadb
langchain-chroma
openai>=1.0.0
python-dotenv
tiktoken
sentence-transformers
fastapi
uvicorn[standard]
pydantic
pydantic-settings
pyodbc
sqlalchemy
